feat: restaurant-detail-content 컴포넌트 생성
